+++ firmware/CHANGELOG.md
@@ -30,6 +30,7 @@ Release template (copy/paste this for new release):
 ## Unreleased
 
 ### Added
+ - Cranking Air Amount / Idle RPM Flare: add hold duration curve to lock the initial value before the Crank-to-Run taper begins
  - ETB bench test button: open throttle 300ms without sensors (#9171)
  - PWM support for fan control #5407
  - Per cylinder knock control individual gain control #9404